🦄 refactor(tests): Address auth
Context: API is now protected by "abilities:api-access" middleware
Using Sanctum::actingAs($user, ['api-access']); instead
$this->actingAs.

Other changes:
- rellocated "generateUserData" test function to facilitate its access to other tests
